requirements.txt
# GUI and visuals
PySide6>=6.5             
Pillow>=10.0.0              

# OCR
pytesseract>=0.3.10        

# PDF to image
PyMuPDF>=1.23.0           

# Image processing
opencv-python>=4.8.0      

# Completion and parsing
openai>=1.30.0             
json5>=0.9.14           

# Data validation
pydantic>=2.6.0

#Env
python-dotenv>=1.0.0

